+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG for Sulu Sales
 
 * dev-develop
 
+    * FEATURE     [General]       Updated for usage with recurring price calculation
     * BUGFIX      [CoreBundle]    Fixed creating independent items in item table caused by 
                                   changes in pricing bundle 
     * ENHANCEMENT [General]       Moved db diagram from sulu/docs to sales